Optimize star rendering every so barely
Render function now only searches for stars to render that have
x-values within the possible range around the player character.
Stars are sorted initially by x-value, and the search for stars
that may be close enough to render (checked still by isOnScreen)
is only performed over half the map. Could be more precise,
but right now performance is not an issue.